% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangling.R
\name{add_count_relative}
\alias{add_count_relative}
\title{Add: relative count}
\usage{
add_count_relative(
  data,
  ...,
  .col_abs = "n_abs",
  .col_rel = "n_rel",
  .round = FALSE,
  .digits = 2,
  .ungroup = TRUE
)
}
\arguments{
\item{data}{[\link{tibble}] Data}

\item{...}{[\link{name}, \link{character} or \link{symbol}] Column names}

\item{.col_abs}{[\link{name}, \link{character} or \link{symbol}] Column name for column with
absolute counts}

\item{.col_rel}{[\link{name}, \link{character} or \link{symbol}] Column name for column with
relative counts}

\item{.round}{[\link{logical}] Round relative count yes/no}

\item{.digits}{[\link{integer}] Number of digits when rounding relative count}

\item{.ungroup}{[\link{logical}] Ungroup data before returning it}
}
\value{

}
\description{
Add: relative count
}
\examples{
tibble::tibble(abs = 1:10) \%>\%
    add_count_relative(.col_abs = "abs", .col_rel = "rel",
        .round = TRUE, .digits = 2)
}
