% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangling.R
\name{format_round}
\alias{format_round}
\title{Format: round}
\usage{
format_round(data, ..., .exec = TRUE, .digits = 0)
}
\arguments{
\item{data}{[\link{tibble}]}

\item{...}{[\link{name}, \link{symbol} or \link{character}] NSE-flexible column name}

\item{.exec}{[\link{logical}] Execute functionality yes/no}

\item{.digits}{[\link{integer}] Number of digits for rounding}
}
\value{

}
\description{
Functional programming wrapper around `\link[dplyr:mutate]{dplyr::mutate}.
}
\examples{
tibble::tribble(
    ~a, ~b, ~c,
    "a", 1.123, 0.12345,
    "b", 1.567, 0.567
) \%>\% format_round(b, c, .digits = 2)
}
