% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nse.R
\name{handle_nse_input}
\alias{handle_nse_input}
\title{Handle NSE input}
\usage{
handle_nse_input(x)
}
\arguments{
\item{x}{[\link{character} or \link{name} or \link{call}]}
}
\value{

}
\description{
Handle NSE input
}
