% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangling.R
\name{struc_ungroup}
\alias{struc_ungroup}
\title{Structure: ungroup}
\usage{
struc_ungroup(data, exec)
}
\arguments{
\item{data}{[\link{tibble}] Data}

\item{exec}{[\link{logical}] Execute inner functionality yes/no}
}
\value{

}
\description{
Functional programming wrapper around \link[dplyr:group_by]{dplyr::ungroup}..
}
\examples{
mtcars \%>\%
    dplyr::group_by(cyl) \%>\%
    struc_ungroup(TRUE)

mtcars \%>\%
    dplyr::group_by(cyl) \%>\%
    struc_ungroup(FALSE)
}
