% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangling.R
\name{summa_freq_table}
\alias{summa_freq_table}
\title{Frequency table}
\usage{
summa_freq_table(
  data,
  ...,
  .sort = FALSE,
  .col_n_abs = "n_abs",
  .col_n_rel = "n_rel",
  .round_n_rel = FALSE,
  .digits_n_rel = 0,
  .ungroup = TRUE
)
}
\arguments{
\item{data}{[\link{tibble}] Data}

\item{...}{[\link{name}, \link{character} or \link{symbol}] Column names}

\item{.sort}{[\link{logical}] Should results be sorted yes/no}

\item{.col_n_abs}{[\link{symbol} or \link{character}] Column name for absolute count}

\item{.col_n_rel}{[\link{symbol} or \link{character}] Column name for relative count}

\item{.round_n_rel}{[\link{logical}] Round \code{n_rel} yes/no}

\item{.digits_n_rel}{[\link{integer}] Number of digits for rounding}

\item{.ungroup}{[\link{logical}] Ungroup data frame yes/no}
}
\value{

}
\description{
Frequency table
}
\examples{
mtcars \%>\% summa_freq_table(cyl)
mtcars \%>\% summa_freq_table(cyl, gear, .sort = TRUE)

mtcars \%>\% summa_freq_table(cyl, .col_n_abs = "n",
    .col_n_rel = rel, .sort = TRUE)

mtcars \%>\% summa_freq_table(cyl, "gear")

cyl_ <- dplyr::quo(cyl)
gear_ <- dplyr::sym("gear")
mtcars \%>\% summa_freq_table(!!cyl_, !!gear_)

library(palmerpenguins)
penguins \%>\% summa_freq_table(species, island, var_dep = sex, .digits_n_rel = 2)

# Check that relative counts add up to `1`
penguins \%>\%
    summa_freq_table(species, island, var_dep = sex,
        .digits_n_rel = 2, .ungroup = FALSE) \%>\%
    dplyr::summarise(n_rel_total = sum(n_rel)) \%>\%
    dplyr::ungroup()
}
