% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangling.R
\name{wr_freq_table}
\alias{wr_freq_table}
\title{Frequency table}
\usage{
wr_freq_table(
  data,
  ...,
  .sort = FALSE,
  .col_n_abs = "n_abs",
  .col_n_rel = "n_rel",
  .digits_n_rel = 0,
  .ungroup = TRUE
)
}
\arguments{
\item{data}{Data}

\item{...}{[\link{symbol}, \link{character}]}

\item{.sort}{[\link{logical}] Should results be sorted yes/no}

\item{.col_n_abs}{[\link{symbol} or \link{character}] Column name for absolute count}

\item{.col_n_rel}{[\link{symbol} or \link{character}] Column name for relative count}

\item{.digits_n_rel}{[\link{integer}] Number of digits for rounding. \code{0} means "no
rounding"}

\item{.ungroup}{[\link{logical}] Ungroup data frame yes/no}

\item{var_dep}{[\link{symbol} or \link{character}]}
}
\value{

}
\description{
Frequency table
}
\examples{
mtcars \%>\% wr_freq_table(cyl)
mtcars \%>\% wr_freq_table(cyl, gear, .sort = TRUE)

mtcars \%>\% wr_freq_table(cyl, .col_n_abs = "n",
    .col_n_rel = rel, .sort = TRUE)

mtcars \%>\% wr_freq_table(cyl, "gear")

cyl_ <- dplyr::quo(cyl)
gear_ <- dplyr::sym("gear")
mtcars \%>\% wr_freq_table(!!cyl_, !!gear_)

library(palmerpenguins)
penguins \%>\% wr_freq_table(species, island, var_dep = sex, .digits_n_rel = 2)

Check that relative counts add up to `1`
penguins \%>\%
    wr_freq_table(species, island, var_dep = sex,
        .digits_n_rel = 2, .ungroup = FALSE) \%>\%
    dplyr::summarise(n_rel_total = sum(n_rel)) \%>\%
    dplyr::ungroup()
}
