% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangling.R
\name{wr_summarize}
\alias{wr_summarize}
\title{Summarize}
\usage{
wr_summarize(.data, ..., .groups = NULL)
}
\arguments{
\item{.groups}{}
}
\value{

}
\description{
Summarize
}
\examples{
mtcars \%>\% wr_summarize(
    carb_sum = sum(carb, na.rm = TRUE),
    carb_mean = mean(carb, na.rm = TRUE)
)
}
