% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrang-package.R
\docType{package}
\name{wrang-package}
\alias{wrang}
\alias{wrang-package}
\title{wrang: Swiss army knife for data wrangling}
\description{
Custom functions based upon 'dplyr' and tidyverse friends to make typical data wrangling tasks easier
}
\author{
\strong{Maintainer}: Janko Thyson \email{janko@rappster.io}

}
\keyword{internal}
