% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangling.R
\name{wr_freq_table}
\alias{wr_freq_table}
\title{Frequency table}
\usage{
wr_freq_table(
  data,
  ...,
  .sort = FALSE,
  .col_n_abs = "n_abs",
  .col_n_rel = "n_rel"
)
}
\arguments{
\item{data}{Data}

\item{...}{[\link{symbol}, \link{character}]}

\item{.sort}{[\link{logical}] Should results be sorted yes/no}

\item{.col_n_abs}{[\link{symbol} or \link{character}] Column name for absolute count}

\item{.col_n_rel}{[\link{symbol} or \link{character}] Column name for relative count}
}
\value{

}
\description{
Frequency table
}
\examples{
mtcars \%>\% wr_freq_table(cyl)
mtcars \%>\% wr_freq_table(cyl, gear, .sort = TRUE)

mtcars \%>\% wr_freq_table(cyl, .col_n_abs = "n",
    .col_n_rel = rel, .sort = TRUE)

mtcars \%>\% wr_freq_table(cyl, "gear")

cyl_ <- dplyr::quo(cyl)
gear_ <- dplyr::sym("gear")
mtcars \%>\% wr_freq_table(!!cyl_, !!gear_)
}
